feat(values): unify EmptyList* — collapse syntax/value duality
Move SourceContext, OriginInfo, SourceIndexes, Scope, scope-set
utilities, SyntaxValue interface, SyntaxTuple interface, and
SyntaxVector into the values package. The empty-list singleton
(values.EmptyList) now satisfies SyntaxValue and SyntaxTuple
directly, so values.SyntaxEmptyList is the same struct value as
values.EmptyList.

The internal/syntax package keeps the concrete syntax data types
(SyntaxObject, SyntaxSymbol, SyntaxPair, SyntaxComment, etc.) and
re-exports the moved types via Go type aliases so existing import
sites compile unchanged. Two hooks (UnwrapAllSharedFunc and
SyntaxVectorAddScopeFunc, plus SyntaxVoidSingleton) bridge methods on
values.SyntaxVector that need to traverse concrete syntax types
defined upstream — internal/syntax/syntax_vector.go registers them
in init().

Behavior changes:
- (equal? (syntax ()) '()) now returns #t (matching Chez 10.3.0)
- IsSyntaxEmptyList delegates to values.IsEmptyList (symmetric)
- The deleted syntax_empty_list.go's strict pointer-type EqualTo is
  gone; the comment claiming this matched Chez/Racket was incorrect.

Files deleted:
- internal/syntax/syntax_empty_list.go
- internal/syntax/syntax_empty_list_test.go
