+++ VERSION
@@ -1,1 +1,1 @@
-v1.15.22
+v1.15.23
