+++ internal/syntax/syntax_value.go
@@ -16,7 +16,6 @@ package syntax
 
 import (
 	"fmt"
-	"sync/atomic"
 
 	"github.com/aalpar/wile/values"
 	"github.com/aalpar/wile/werr"
@@ -27,74 +26,24 @@ var (
 	_ SyntaxValue  = (*SyntaxObject)(nil)
 )
 
-// Scope is an identity marker for macro hygiene.
-// Each macro invocation creates a fresh Scope. Hygiene checking uses pointer
-// equality to determine if a binding's scopes are a subset of a reference's scopes.
-// This implements Flatt's "sets of scopes" model where scopes are just unique tags,
-// not environment hierarchies.
-type Scope struct {
-	id uint64 // ensures unique pointer identity (empty structs can share addresses in Go)
-	// IsRebinding indicates whether this scope can potentially rebind auxiliary syntax.
-	// True for let-syntax/letrec-syntax scopes which create local macro bindings.
-	// False for with-binding-scope which only adds scopes for binding hygiene.
-	// This distinction is used in literalScopesMatch to correctly handle auxiliary
-	// syntax like => and else in cond/case.
-	IsRebinding bool
-	// Label is an optional human-readable description for debugging.
-	// Examples: "lambda", "let-syntax", "intro:my-macro", "library:(wile kanren)".
-	Label string
-}
-
-// nextScopeID is a counter for generating unique scope identities
-var nextScopeID atomic.Uint64
+// Scope is the macro-hygiene identity marker. Defined in package values
+// alongside SourceContext (the empty-list duality merge).
+type Scope = values.Scope
 
 // NewScope creates a new scope with unique identity for hygiene tracking.
-// By default, scopes are not rebinding scopes.
-func NewScope() *Scope {
-	id := nextScopeID.Add(1)
-	return &Scope{id: id}
-}
+var NewScope = values.NewScope
 
 // NewScopeWithLabel creates a new scope with a human-readable label for debugging.
-// The label has no semantic effect — it is purely for diagnostics.
-func NewScopeWithLabel(label string) *Scope {
-	id := nextScopeID.Add(1)
-	return &Scope{id: id, Label: label}
-}
+var NewScopeWithLabel = values.NewScopeWithLabel
 
 // NewRebindingScope creates a new scope that can potentially rebind auxiliary syntax.
-// Used by let-syntax and letrec-syntax to mark scopes that could shadow literals.
-func NewRebindingScope() *Scope {
-	id := nextScopeID.Add(1)
-	return &Scope{id: id, IsRebinding: true}
-}
+var NewRebindingScope = values.NewRebindingScope
 
 // NewRebindingScopeWithLabel creates a new rebinding scope with a label.
-func NewRebindingScopeWithLabel(label string) *Scope {
-	id := nextScopeID.Add(1)
-	return &Scope{id: id, IsRebinding: true, Label: label}
-}
+var NewRebindingScopeWithLabel = values.NewRebindingScopeWithLabel
 
-// ID returns the unique identifier for this scope.
-// This can be used as a macro application ID for tracing.
-func (p *Scope) ID() uint64 {
-	if p == nil {
-		return 0
-	}
-	return p.id
-}
-
-// String returns a human-readable representation of the scope.
-// If a label is set, returns "scope:ID(label)"; otherwise "scope:ID".
-func (p *Scope) String() string {
-	if p == nil {
-		return "scope:nil"
-	}
-	if p.Label != "" {
-		return fmt.Sprintf("scope:%d(%s)", p.id, p.Label)
-	}
-	return fmt.Sprintf("scope:%d", p.id)
-}
+// (Scope.ID and Scope.String are defined on values.Scope — see values/scope.go.)
+var _ = fmt.Sprintf // keep fmt import for SchemeString below
 
 // syntaxBase provides common SourceContext() implementation for syntax types.
 // This embedded struct eliminates boilerplate SourceContext() methods across 9 syntax types.
