+++ internal/syntax/syntax_vector.go
@@ -14,149 +14,41 @@
 
 package syntax
 
-import (
-	"context"
-	"strings"
-
-	"github.com/aalpar/wile/values"
-)
-
-var (
-	_ values.Value = (*SyntaxVector)(nil)
-	_ SyntaxValue  = (*SyntaxVector)(nil)
-)
+import "github.com/aalpar/wile/values"
 
 // SyntaxVector wraps a Scheme vector with source context.
-type SyntaxVector struct {
-	Values []SyntaxValue
-	syntaxBase
-}
-
-// AddScope recursively propagates a scope to all nested syntax values.
-//
-// This implements scope propagation for Flatt's "sets of scopes" hygiene.
-// When a macro expands, the intro scope must be added to all identifiers
-// (symbols) in the expansion. This method walks the vector structure and
-// calls AddScope on each element, ultimately reaching the symbols.
-//
-// Only symbols store scopes for hygiene resolution. Vectors just propagate.
-func (p *SyntaxVector) AddScope(scope *Scope) SyntaxValue {
-	// Nil or empty vector has no symbols to propagate to
-	if p == nil || len(p.Values) == 0 {
-		return p
-	}
-
-	// Use the generic mapSyntaxTree traversal to add scope to all nested nodes.
-	// mapSyntaxTree handles vector recursion and calls the function on elements.
-	return mapSyntaxTree(p, func(node SyntaxValue) SyntaxValue {
-		// Try to call AddScope on nodes that support it (symbols, pairs, vectors, etc.)
-		adder, ok := node.(interface{ AddScope(*Scope) SyntaxValue })
-		if ok {
-			return adder.AddScope(scope)
-		}
-		// Nodes that don't support AddScope are returned unchanged (SyntaxObject)
-		return node
-	})
-}
+// Defined in package values (the empty-list duality merge — see
+// values/syntax_vector.go). Methods that need to traverse concrete
+// syntax types defined in this package (AddScope, UnwrapAll) dispatch
+// via hooks registered at init time below.
+type SyntaxVector = values.SyntaxVector
 
 // NewSyntaxVector creates a new syntax vector with the given source context and elements.
-func NewSyntaxVector(sc *SourceContext, vs ...SyntaxValue) *SyntaxVector {
-	q := &SyntaxVector{
-		Values: vs,
-		syntaxBase: syntaxBase{
-			sourceContext: sc,
-		},
-	}
-	return q
-}
+var NewSyntaxVector = values.NewSyntaxVector
 
-// UnwrapAll recursively unwraps all elements to produce a plain values.Vector.
-func (p *SyntaxVector) UnwrapAll() values.Value {
-	return UnwrapAllShared(p, make(map[SyntaxValue]values.Value))
-}
+func init() {
+	// Recursive scope propagation across nested syntax types lives here in
+	// internal/syntax (where the concrete syntax types are defined). The
+	// values.SyntaxVector.AddScope method dispatches to this hook.
+	values.SyntaxVectorAddScopeFunc = func(p *SyntaxVector, scope *Scope) values.SyntaxValue {
+		return mapSyntaxTree(p, func(node values.SyntaxValue) values.SyntaxValue {
+			adder, ok := node.(interface {
+				AddScope(*Scope) values.SyntaxValue
+			})
+			if ok {
+				return adder.AddScope(scope)
+			}
+			return node
+		})
+	}
 
-func (p *SyntaxVector) Unwrap() values.Value {
-	if p.IsVoid() {
-		return values.Void
-	}
-	vq := make([]values.Value, len(p.Values))
-	for i, v := range p.Values {
-		vq[i] = v
-	}
-	q := values.NewVector(vq...)
-	return q
-}
+	// Cycle-aware recursive unwrap. The implementation below dispatches on
+	// concrete syntax types defined in this package and is registered into
+	// values so that values.SyntaxVector.UnwrapAll can call it.
+	values.UnwrapAllSharedFunc = UnwrapAllShared
 
-// IsVoid returns true if the syntax vector is nil.
-func (p *SyntaxVector) IsVoid() bool {
-	return p == nil
-}
-
-// SchemeString returns the Scheme representation of the syntax vector.
-func (p *SyntaxVector) SchemeString() string {
-	if p.IsVoid() {
-		return "#'<void>"
-	}
-	q := strings.Builder{}
-	q.WriteString("#'(")
-	for i, v := range p.Values {
-		if i > 0 {
-			q.WriteString(" ")
-		}
-		q.WriteString(v.SchemeString())
-	}
-	q.WriteString(")")
-	return q.String()
-}
-
-// EqualTo performs pointer comparison only, matching Chez Scheme/Racket behavior.
-// Two syntax objects are equal? only if they are the same object.
-// For value comparison of syntax objects, use bound-identifier=? or free-identifier=?.
-func (p *SyntaxVector) EqualTo(o values.Value) bool {
-	v, ok := o.(*SyntaxVector)
-	if !ok {
-		return false
-	}
-	return p == v
-}
-
-// ForEach iterates over the elements of the vector as regular values in index order.
-// It provides tuple-style iteration compatible with values.ForEachFunc callbacks.
-func (p *SyntaxVector) ForEach(ctx context.Context, fn values.ForEachFunc) (values.Value, error) {
-	if p.IsVoid() {
-		return values.Void, nil
-	}
-	for i, v := range p.Values {
-		hasNext := i+1 < len(p.Values)
-		err := fn(ctx, i, hasNext, v)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return values.EmptyList, nil
-}
-
-// SyntaxForEach iterates over the syntax elements of the vector.
-// A nil receiver is treated as the distinguished syntax void value and
-// results in no iteration and a SyntaxVoid tail.
-//
-// The callback is invoked for each element with its index and a boolean
-// indicating whether there is another element after the current one.
-// If the callback returns an error, iteration stops immediately and the
-// error is returned.
-func (p *SyntaxVector) SyntaxForEach(ctx context.Context, fn SyntaxForEachFunc) (SyntaxValue, error) {
-	if p.IsVoid() {
-		return SyntaxVoid, nil
-	}
-	for i, v := range p.Values {
-		hasNext := i+1 < len(p.Values)
-		err := fn(ctx, i, hasNext, v)
-		if err != nil {
-			return nil, err
-		}
-	}
-	// Vectors do not have a list tail, so we return SyntaxEmptyList as a
-	// conventional "no remainder" sentinel for callers that expect a tail
-	// value similar to SyntaxPair.SyntaxForEach.
-	return SyntaxEmptyList, nil
+	// Register the syntax-level void singleton so values.SyntaxVector with
+	// a nil receiver returns it from SyntaxForEach (preserves the original
+	// "syntax void tail" behavior).
+	values.SyntaxVoidSingleton = SyntaxVoid
 }
